restructure language to modules
