+++ editor/highlight.py
